+++ CHANGELOG.md
@@ -21,6 +21,12 @@ These changes are available on the `master` branch, but have not yet been releas
 
 ### Fixed
 
+- Fix `StageChannel.create_instance`'s `send_notification` parameter type.
+  ([#3273](https://github.com/Pycord-Development/pycord/pull/3273))
+- Fix the `sync_permissions` type in the `VoiceChannel.edit` and `StageChannel.edit`
+  overloads. ([#3273](https://github.com/Pycord-Development/pycord/pull/3273))
+- Fix the return type of the `MediaChannel.edit` overload.
+  ([#3273](https://github.com/Pycord-Development/pycord/pull/3273))
 - Fix `TypeError` when accessing `ApplicationCommand.guild_only` or
   `SlashCommandGroup.guild_only` when `contexts` is `None`.
   ([#3320](https://github.com/Pycord-Development/pycord/pull/3320))